Fix full package stubtest
This commit fixes most of the package for stubtest to work. This
primarily involves having the stubfiles match the package layout.
Unfortunately this means we need to put the majority of the type hints
in the stubfile for `rustworkx.rustworkx`. Having the stubfiles split up
as before mypy complains that the module doesn't exist. For example,
having the coloring functions in `rustworkx/coloring.pyi` causes a mypy
error that `rustworkx.coloring` couldn't be imported. This causes the
rustworkx.pyi file to be quite large, but I couldn't find a pattern
to workaround this limitation. Aside from that a bunch of stubs needed
to be updated as now that they're getting included in the stubtest run
issues were caught. The one exception is the generators module where the
custom `text_signature` set in the function definitions were incorrect
and tripping up mypy, these were fixed by removing the `text_signature`
fields.
