--- rustworkx/coloring.pyi
@@ -1,21 +1,0 @@
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-
-# This file contains only type annotations for PyO3 functions and classes
-# For implementation details, see __init__.py and src/coloring.rs
-
-from .graph import PyGraph
-from .digraph import PyDiGraph
-
-def graph_greedy_color(graph: PyGraph, /) -> dict[int, int]: ...
-def graph_greedy_edge_color(graph: PyGraph, /) -> dict[int, int]: ...
-def graph_is_bipartite(graph: PyGraph) -> bool: ...
-def digraph_is_bipartite(graph: PyDiGraph) -> bool: ...
-def graph_two_color(graph: PyGraph) -> dict[int, int]: ...
-def digraph_two_color(graph: PyDiGraph) -> dict[int, int]: ...
-def graph_misra_gries_edge_color(graph: PyGraph, /) -> dict[int, int]: ...
