--- rustworkx/union.pyi
@@ -1,33 +1,0 @@
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-
-# This file contains only type annotations for PyO3 functions and classes
-# For implementation details, see __init__.py and src/union.rs
-
-from .graph import PyGraph
-from .digraph import PyDiGraph
-
-from typing import TypeVar
-
-_S = TypeVar("_S")
-_T = TypeVar("_T")
-
-def digraph_union(
-    first: PyDiGraph[_S, _T],
-    second: PyDiGraph[_S, _T],
-    /,
-    merge_nodes: bool = ...,
-    merge_edges: bool = ...,
-) -> PyDiGraph[_S, _T]: ...
-def graph_union(
-    first: PyGraph[_S, _T],
-    second: PyGraph[_S, _T],
-    /,
-    merge_nodes: bool = ...,
-    merge_edges: bool = ...,
-) -> PyGraph[_S, _T]: ...
