+++ rustworkx/visit.pyi
@@ -34,6 +34,6 @@ class DFSVisitor(Generic[_T]):
 class DijkstraVisitor(Generic[_T]):
     def discover_vertex(self, v: int, score: float): ...
     def finish_vertex(self, v: int): ...
-    def examine_edge(self, e: tuple[int, int, _T]): ...
-    def edge_relaxed(self, e: tuple[int, int, _T]): ...
-    def edge_not_relaxed(self, e: tuple[int, int, _T]): ...
+    def examine_edge(self, edge: tuple[int, int, _T]): ...
+    def edge_relaxed(self, edge: tuple[int, int, _T]): ...
+    def edge_not_relaxed(self, edge: tuple[int, int, _T]): ...
